+++ ProjetoPI2/src/view/CadClientView.form
@@ -49,7 +49,7 @@
     <Property name="title" type="java.lang.String" value="Cadastro Cliente"/>
     <Property name="alwaysOnTop" type="boolean" value="true"/>
     <Property name="cursor" type="java.awt.Cursor" editor="org.netbeans.modules.form.editors2.CursorEditor">
-      <Color id="Cursor Padr&#xe3;o"/>
+      <Color id="Default Cursor"/>
     </Property>
     <Property name="resizable" type="boolean" value="false"/>
   </Properties>
@@ -249,7 +249,7 @@
                 </Property>
                 <Property name="toolTipText" type="java.lang.String" value="Sexo"/>
                 <Property name="cursor" type="java.awt.Cursor" editor="org.netbeans.modules.form.editors2.CursorEditor">
-                  <Color id="Cursor de Seta Que Aponta &#xe0; Direita"/>
+                  <Color id="Default Cursor"/>
                 </Property>
               </Properties>
               <AuxValues>
