+++ ProjetoPI2/src/view/Relatorio.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Pacote_relatorio;
+package view;
 
 /**
  *
